requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-ratelimit==4.1.0
django-extensions==3.2.3
cryptography==41.0.7
pyjwt==2.8.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
Pillow==10.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
django-extensions==3.2.3
python-decouple==3.8
django-filter==23.5
django-crispy-forms==2.1
crispy-bootstrap4==2022.1
psutil==5.9.8
django-health-check==3.17.0

# Testing dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
factory-boy==3.3.0
faker==22.0.0
freezegun==1.4.0

# API Documentation
drf-spectacular==0.27.0